+++ boot/core/src/boot/task/built_in.clj
@@ -34,7 +34,8 @@
     (let [tasks (#'helpers/available-tasks 'boot.user)
           opts  (->> main/cli-opts (mapv (fn [[x y z]] ["" (str x " " y) z])))
           envs  [["" "BOOT_AS_ROOT"              "Set to 'yes' to allow boot to run as root."]
-                 ["" "BOOT_CLOJARS_MIRROR"       "Specify mirror for clojars.org for boot's own dependencies."]
+                 ["" "BOOT_CLOJARS_REPO"         "Specify the url for the 'clojars' Maven repo."]
+                 ["" "BOOT_CLOJARS_MIRROR"       "Specify the mirror url for the 'clojars' Maven repo."]
                  ["" "BOOT_CLOJURE_VERSION"      "The version of Clojure boot will provide (1.7.0)."]
                  ["" "BOOT_CLOJURE_NAME"         "The artifact name of Clojure boot will provide (org.clojure/clojure)."]
                  ["" "BOOT_COLOR"                "Set to 'no' to turn colorized output off."]
@@ -45,10 +46,12 @@
                  ["" "BOOT_JAVA_COMMAND"         "Specify the Java executable (java)."]
                  ["" "BOOT_JVM_OPTIONS"          "Specify JVM options (Unix/Linux/OSX only)."]
                  ["" "BOOT_LOCAL_REPO"           "The local Maven repo path (~/.m2/repository)."]
-                 ["" "BOOT_MAVEN_CENTRAL_MIRROR" "Specify mirror for repo1.maven.org for boot's own dependencies."]
+                 ["" "BOOT_MAVEN_CENTRAL_REPO"   "Specify the url for the 'maven-central' Maven repo."]
+                 ["" "BOOT_MAVEN_CENTRAL_MIRROR" "Specify the mirror url for the 'maven-central' Maven repo."]
                  ["" "BOOT_VERSION"              "Specify the version of boot core to use."]
                  ["" "BOOT_WARN_DEPRECATED"      "Set to 'no' to suppress deprecation warnings."]]
           files [["" "./boot.properties"         "Specify boot options for this project."]
+                 ["" "./profile.boot"            "A script to run after the global profile.boot but before the build script."]
                  ["" "BOOT_HOME/boot.properties" "Specify global boot options."]
                  ["" "BOOT_HOME/profile.boot"    "A script to run before running the build script."]]
           br    #(conj % ["" "" ""])]